fundooNotes-master/requirements.txt
coverage==4.4.1
psycopg2==2.7.3
Django==2.0.1
django-nose==1.4.5
djangorestframework==3.7.7
nose==1.3.7
